app/generator_service/requirements.txt
python-dotenv
psycopg2-binary
numpy
sentence-transformers
httpx
openai
fastapi
uvicorn
